refactor(core): nest overlay-code-default provider under core/overlays
The #36 provider was added as a new root leaf under teatree.core
(overlay_code_defaults_provider.py), pushing the flat-core-leaf ratchet
to 68 vs the pinned 67 (tests/quality/test_no_flat_core_regrowth.py).

Move it into a new teatree.core.overlays package — the core-side home for
overlay seams whose registration point lives below in teatree.config. The
count returns to 67 without touching the pin. overlay_loader.py already sits
at the module-health public-function ceiling (10) and overlay.py is over the
LOC ceiling, so the seam stays its own small file rather than inlining. The
inverted dependency is unchanged: the provider registers into
teatree.config.overlay_code_defaults at import time via a dependency-injected
get_overlay, so the platform layer still never imports the domain
(tach + lint-imports green). The test mirrors the module at
tests/teatree_core/overlays/.

Give the lifecycle-probe test mocks a real UserSettings for config.user: the
new overlay-code-default tier makes resolution apply the promoted keys via
dataclasses.replace even with no DB overrides, which the prior MagicMock stub
could not satisfy.
